Chrome Radio V1.0
This is the official repo for Chrome Radio. As this is the first version, expect a ton of bugs.
You can check the "common-bugs.txt" file to check some common bugs we are aware of and trying to fix.
If you find a bug that isn't in the "common-bugs.txt" file, please report it using the "Issues" tab in Github.

Please support this project as I'm a 13-year-old developer doing all this by myself.